+++ packages/terra-date-picker/src/react-datepicker/calendar.jsx
@@ -32,7 +32,8 @@ import {
   allDaysDisabledAfter,
   getEffectiveMinDate,
   getEffectiveMaxDate,
-  isDayDisabled
+  isDayDisabled,
+  dateValues
 } from './date_utils'
 
 const cx = classNames.bind(styles);
@@ -413,7 +414,7 @@ export default class Calendar extends React.Component {
       isMonthChanged: true,
       date: getStartOfMonth(addMonths(cloneDate(this.state.date), 1)),
     }, () => this.handleMonthChange(this.state.date))
-    this.props.setPreSelection(getStartOfMonth(addMonths(cloneDate(this.state.date), 1)));
+    this.props.setPreSelection(getStartOfMonth(addMonths(cloneDate(this.state.date), 1)),dateValues.MONTH,addMonths(cloneDate(this.state.date), 1));
     // To check if button is pressed using mouse or keyboard
     if(event.target.type === undefined) {
       this.setState({ calendarIsKeyboardFocused : false});
@@ -426,7 +427,7 @@ export default class Calendar extends React.Component {
       isMonthChanged: true,
       date: getStartOfMonth(subtractMonths(cloneDate(this.state.date), 1))
     }, () => this.handleMonthChange(this.state.date))
-    this.props.setPreSelection(getStartOfMonth(subtractMonths(cloneDate(this.state.date), 1)));
+    this.props.setPreSelection(getStartOfMonth(subtractMonths(cloneDate(this.state.date), 1)),dateValues.MONTH,subtractMonths(cloneDate(this.state.date), 1));
     // To check if button is pressed using mouse or keyboard
     if(event.target.type === undefined) {
       this.setState({ calendarIsKeyboardFocused : false});
@@ -464,7 +465,7 @@ export default class Calendar extends React.Component {
       isMonthChanged: true,
       date: getStartOfMonth(setYear(cloneDate(this.state.date), year))
     })
-    this.props.setPreSelection(getStartOfMonth(setYear(cloneDate(this.state.date), year)));
+    this.props.setPreSelection(getStartOfMonth(setYear(cloneDate(this.state.date), year)),dateValues.YEAR,year);
   }
 
   changeMonth = (month) => {
@@ -472,7 +473,7 @@ export default class Calendar extends React.Component {
       isMonthChanged: true,
       date: getStartOfMonth(setMonth(cloneDate(this.state.date), month))
     }, () => this.handleMonthChange(this.state.date))
-    this.props.setPreSelection(getStartOfMonth(setMonth(cloneDate(this.state.date), month)));
+    this.props.setPreSelection(getStartOfMonth(setMonth(cloneDate(this.state.date), month)),dateValues.MONTH,month);
   }
 
   header = (date = this.state.date) => {