+++ packages/terra-date-picker/src/DatePicker.jsx
@@ -141,6 +141,12 @@ const propTypes = {
    * The value must be in the `YYYY-MM-DD` format or the all-numeric date format based on the locale.
    */
   value: PropTypes.string,
+  /**
+   * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue).
+   * If invalid error text is used, provide a string containing the IDs for error html element.
+   * ID must be htmlFor prop value with error text.
+   */
+  errorId: PropTypes.string,
 };
 
 const defaultProps = {
@@ -165,6 +171,7 @@ const defaultProps = {
   useExternalFormatMask: false,
   required: false,
   selectedDate: undefined,
+  errorId: '',
 };
 
 class DatePicker extends React.Component {
@@ -427,6 +434,7 @@ class DatePicker extends React.Component {
       selectedDate,
       useExternalFormatMask,
       value,
+      errorId,
       ...customProps
     } = this.props;
 
@@ -506,6 +514,7 @@ class DatePicker extends React.Component {
                 maxDate={maxDate}
                 minDate={minDate}
                 filterDate={this.handleFilterDate}
+                errorId={errorId}
               />
             )}
             customInputRef="firstInputRefCallback"