+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Changed
+  * Fixed the invalid case for SR announcement.
+  * Changed the SR announcement for changing date from calender.
+
 ## 4.98.1 - (February 16, 2024)
 
 * Fixed