+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Added
+  * Added visual focus dashed border for `terra-date-picker`.
+  * Removed selected date information on focus of picker button.
+  * Added ariaDescribedBy prop in `terra-date-time-picker` example for invalid error message.
+
 * Changed
   * Fixed the invalid case for SR announcement.
   * Changed the SR announcement for changing date from calender.